requirements.txt
numpy
Pillow
matplotlib
pyyaml
argparse